Add ServiceClientMock
